+++ packages/node/src/client.ts
@@ -128,6 +128,7 @@ export abstract class PostHogBackendClient extends PostHogCoreStateless implemen
           },
           customHeaders: this.getCustomHeaders(),
           cacheProvider: options.flagDefinitionCacheProvider,
+          strictLocalEvaluation: options.strictLocalEvaluation,
         })
       }
     }
@@ -675,7 +676,7 @@ export abstract class PostHogBackendClient extends PostHogCoreStateless implemen
 
     // set defaults
     if (onlyEvaluateLocally == undefined) {
-      onlyEvaluateLocally = false
+      onlyEvaluateLocally = this.options.strictLocalEvaluation ?? false
     }
     if (sendFeatureFlagEvents == undefined) {
       sendFeatureFlagEvents = this.options.sendFeatureFlagEvent ?? true
@@ -883,7 +884,7 @@ export abstract class PostHogBackendClient extends PostHogCoreStateless implemen
 
     // set defaults
     if (onlyEvaluateLocally == undefined) {
-      onlyEvaluateLocally = false
+      onlyEvaluateLocally = this.options.strictLocalEvaluation ?? false
     }
 
     const payloadWasLocallyEvaluated = response !== undefined
@@ -1104,7 +1105,7 @@ export abstract class PostHogBackendClient extends PostHogCoreStateless implemen
 
     // set defaults
     if (onlyEvaluateLocally == undefined) {
-      onlyEvaluateLocally = false
+      onlyEvaluateLocally = this.options.strictLocalEvaluation ?? false
     }
 
     const localEvaluationResult = await this.featureFlagsPoller?.getAllFlagsAndPayloads(
@@ -1491,7 +1492,8 @@ export abstract class PostHogBackendClient extends PostHogCoreStateless implemen
     const flagKeys = sendFeatureFlagsOptions?.flagKeys
 
     // Check if we should only evaluate locally
-    const onlyEvaluateLocally = sendFeatureFlagsOptions?.onlyEvaluateLocally ?? false
+    const onlyEvaluateLocally =
+      sendFeatureFlagsOptions?.onlyEvaluateLocally ?? this.options.strictLocalEvaluation ?? false
 
     // If onlyEvaluateLocally is true, only use local evaluation
     if (onlyEvaluateLocally) {